Update 7.1-exception.ipynb
In the very third cell from bottom u typed <file.closed()> but <file.close>
here there won't be any parenthesis as closed is not a method but a property which returns boolean value. And its not throwing error bcoz it is not getting executed as u have used or operator.